feat: sync benchmark chart tooltip theme
